@@ -23,7 +23,6 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
@@ -36,7 +35,7 @@ classifiers = [
     "Topic :: Text Processing :: Filters",
     "Topic :: Utilities",
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 dependencies = [
     "python-dateutil>=2.8.0",
     "colorama>=0.4.0",
@@ -72,7 +71,7 @@ loggrep = ["py.typed"]
 # Development tools configuration
 [tool.black]
 line-length = 88
-target-version = ['py37']
+target-version = ['py38']
 
 [tool.isort]
 profile = "black"
