@@ -5,17 +5,13 @@ Contains the main LogSearcher class that handles pattern matching,
 timestamp filtering, and context line management.
 """
 
+import io
 import re
 import sys
 from collections import deque
-from datetime import datetime
-from typing import Dict, Iterator, List, Optional, TextIO, Union
+from typing import Iterator, List, Optional, TextIO
 
-from .timestamps import (
-    detect_timestamp_format,
-    detect_timestamp_format_cached,
-    parse_timestamp,
-)
+from .timestamps import detect_timestamp_format, parse_timestamp
 
 try:
     from colorama import Fore, Style, init
@@ -24,6 +20,8 @@ try:
     COLOR_AVAILABLE = True
 except ImportError:
     COLOR_AVAILABLE = False
+
+NO_TIMESTAMP_THRESHOLD = 3
 
 
 class LogSearcher:
@@ -40,189 +38,101 @@ class LogSearcher:
         use_color: bool = False,
         startup_time: Optional[str] = None,
     ):
-        """Initialize the LogSearcher.
-
-        Args:
-            patterns: List of regex patterns to search for
-            ignore_case: Whether to ignore case in pattern matching
-            invert_match: Whether to invert the match (show non-matching lines)
-            before_context: Number of lines to show before matches
-            after_context: Number of lines to show after matches
-            context: Number of lines to show around matches (overrides before/after)
-            use_color: Whether to use colored output
-            startup_time: Only show matches after this timestamp
-        """
         self.invert_match = invert_match
         self.before_context = max(before_context, context)
         self.after_context = max(after_context, context)
         self.use_color = use_color and COLOR_AVAILABLE
         self.startup_time = parse_timestamp(startup_time) if startup_time else None
 
-        # Compile regex pattern(s) with optimization
         flags = re.IGNORECASE if ignore_case else 0
         try:
-            # Pre-compile pattern for better performance
             if len(patterns) == 1:
-                # Single pattern optimization
                 self.pattern = re.compile(patterns[0], flags)
             else:
-                # Multiple patterns - use alternation
-                self.pattern = re.compile(
-                    "|".join(f"({pattern})" for pattern in patterns), flags
-                )
+                self.pattern = re.compile("|".join(f"({p})" for p in patterns), flags)
         except re.error as e:
             raise ValueError(f"Invalid regex pattern: {e}")
 
-        # Performance optimization: pre-compile common regex for timestamp detection
-        self._timestamp_cache: Dict[str, Optional[str]] = {}
-        self._use_timestamp_filtering = startup_time is not None
-
     def highlight_match(self, line: str, match: re.Match) -> str:
-        """Highlight the matched part of the line with color.
-
-        Args:
-            line: The line containing the match
-            match: The regex match object
-
-        Returns:
-            Line with highlighted match, or original line if no color
-        """
+        """Highlight the matched part of the line with color."""
         if not self.use_color or not COLOR_AVAILABLE:
             return line
         return line.replace(
             match.group(), f"{Fore.RED}{match.group()}{Style.RESET_ALL}"
         )
 
-    def search_stream(
-        self, input_stream: TextIO, use_current_time_default: bool = False
-    ) -> Iterator[str]:
+    def search_stream(self, input_stream: TextIO) -> Iterator[str]:
         """Search through a stream of log lines.
 
-        Args:
-            input_stream: Input stream to search through
-            use_current_time_default: If True and no startup_time specified, use current time
-
-        Yields:
-            Lines that match the search criteria with context
+        Yields lines that match the search criteria, respecting
+        timestamp filtering and context line settings.
         """
-        # Handle context lines
         before_buffer: deque = deque(
             maxlen=self.before_context if self.before_context > 0 else None
         )
         after_count = 0
-        first_timestamp = None
         startup_time = self.startup_time
+        in_range = not startup_time
+        no_timestamp_lines = 0
+        has_seen_timestamp = False
 
-        # If no startup time specified and we're reading from stdin (streaming), use current time
-        # But only if the input appears to be live/streaming (i.e., we detect it's potentially real-time)
-        if not startup_time and use_current_time_default:
-            # For stdin, we'll be more conservative and only use current time if explicitly needed
-            # The caller can control this behavior
-            startup_time = datetime.now()
-
-        # Performance optimization: pre-process lines for better memory handling
-        in_range = not startup_time  # If no startup time, process all lines
-
-        # Try to avoid loading all lines into memory for large files
-        lines: Union[List[str], TextIO]
-        try:
-            if hasattr(input_stream, "seek"):
-                # File-like object, try to get size
-                current_pos = input_stream.tell()
-                input_stream.seek(0, 2)  # Seek to end
-                file_size = input_stream.tell()
-                input_stream.seek(current_pos)  # Restore position
-
-                # For large files (>100MB), use streaming mode
-                if file_size > 100 * 1024 * 1024:  # 100MB
-                    lines = input_stream
-                else:
-                    lines = list(input_stream)
-            else:
-                # stdin or other stream, read all lines
-                lines = list(input_stream)
-        except (AttributeError, OSError):
-            # Fallback to reading all lines
-            lines = list(input_stream)
-
-        for i, line in enumerate(lines):
-            # Performance optimization: only detect timestamps if we need them
-            if startup_time or not in_range:
-                ts_str = (
-                    detect_timestamp_format_cached(line)
-                    if i % 100 == 0
-                    else detect_timestamp_format(line)
-                )
-                ts = None
-
-                # Parse timestamp if found
+        for line in input_stream:
+            # Timestamp filtering
+            if startup_time:
+                ts_str = detect_timestamp_format(line)
                 if ts_str:
+                    has_seen_timestamp = True
                     ts = parse_timestamp(ts_str)
-                    if ts and not first_timestamp:
-                        first_timestamp = ts
-                # Use first timestamp as startup time if none specified and not using current time
-                if (
-                    not startup_time
-                    and not use_current_time_default
-                    and first_timestamp
-                ):
-                    startup_time = first_timestamp
-                    in_range = True
-                # Check if we're past startup time
-                if startup_time and ts and ts >= startup_time:
-                    in_range = True
-                elif startup_time and ts and ts < startup_time:
-                    in_range = False
+                    if ts:
+                        in_range = ts >= startup_time
+                else:
+                    no_timestamp_lines += 1
+                    if (
+                        no_timestamp_lines >= NO_TIMESTAMP_THRESHOLD
+                        and not has_seen_timestamp
+                    ):
+                        # No timestamps found — disable filtering
+                        in_range = True
+                        startup_time = None
 
-            # Process line if we're in the time range or no timestamp filtering
-            if in_range or not startup_time:
-                match = self.pattern.search(line)
-                is_match = (match and not self.invert_match) or (
-                    not match and self.invert_match
-                )
+            if not in_range:
+                continue
 
-                # Handle after-context from previous matches
-                if after_count > 0:
-                    yield line
-                    after_count -= 1
+            match = self.pattern.search(line)
+            is_match = bool(match) != self.invert_match
 
-                if is_match:
-                    # Yield before-context lines
-                    if self.before_context:
-                        context_lines = self.before_context
-                        for before_line in list(before_buffer)[-context_lines:]:
-                            yield before_line
+            # Yield after-context from previous matches
+            if after_count > 0:
+                yield line
+                after_count -= 1
 
-                    # Yield the match line (with highlighting if applicable)
-                    if match:
-                        yield self.highlight_match(line, match)
-                    else:
-                        yield line
-
-                    # Set up after-context
-                    if self.after_context:
-                        after_count = self.after_context
-
-                # Always buffer lines for potential before-context
+            if is_match:
+                # Yield before-context lines
                 if self.before_context:
-                    before_buffer.append(line)
+                    for before_line in before_buffer:
+                        yield before_line
+
+                # Yield the match line (with highlighting if applicable)
+                if match:
+                    yield self.highlight_match(line, match)
+                else:
+                    yield line
+
+                if self.after_context:
+                    after_count = self.after_context
+
+            # Buffer lines for potential before-context
+            if self.before_context:
+                before_buffer.append(line)
 
     def search_file(self, file_path: str) -> Iterator[str]:
-        """Search through a log file.
-
-        Args:
-            file_path: Path to the log file
-
-        Yields:
-            Lines that match the search criteria with context
-        """
+        """Search through a log file."""
         with open(file_path, "r", encoding="utf-8", errors="replace") as f:
-            yield from self.search_stream(f, use_current_time_default=False)
+            yield from self.search_stream(f)
 
     def search_stdin(self) -> Iterator[str]:
-        """Search through stdin.
-
-        Yields:
-            Lines that match the search criteria with context
-        """
-        yield from self.search_stream(sys.stdin, use_current_time_default=False)
+        """Search through stdin with error-tolerant encoding."""
+        stdin_stream = io.TextIOWrapper(
+            sys.stdin.buffer, encoding="utf-8", errors="replace"
+        )
+        yield from self.search_stream(stdin_stream)
