@@ -29,54 +29,102 @@ import pytest
 TEST_DIR = Path(__file__).parent.parent
 LOGGREP_PATH = TEST_DIR / "loggrep"  # Use our console script
 
-# Sample log data for testing different timestamp formats
-SAMPLE_LOG_UNIX_SYSLOG = """Oct  4 11:59:59 server1 service[1234]: Starting service
-Oct  4 12:00:01 server1 service[1234]: Service started successfully
-Oct  4 12:00:05 server1 service[1234]: Processing request 1
-Oct  4 12:00:10 server1 service[1234]: ERROR: Database connection failed
-Oct  4 12:00:12 server1 service[1234]: WARN: Retrying database connection
-Oct  4 12:00:15 server1 service[1234]: INFO: Database connection restored
-Oct  4 12:00:20 server1 service[1234]: Processing request 2
-Oct  4 12:00:25 server1 service[1234]: Request 2 completed successfully
-Oct  4 12:00:30 server1 service[1234]: ERROR: Memory usage high
-Oct  4 12:01:00 server1 service[1234]: Service stopped
-"""
 
-SAMPLE_LOG_ISO8601 = """2025-10-04 11:59:59.123 [INFO] Application starting
-2025-10-04 12:00:01.456 [INFO] Configuration loaded
-2025-10-04 12:00:05.789 [DEBUG] Processing user request
-2025-10-04 12:00:10.012 [ERROR] Database connection timeout
-2025-10-04 12:00:12.345 [WARN] Retrying operation
-2025-10-04 12:00:15.678 [INFO] Operation completed
-2025-10-04 12:00:20.901 [DEBUG] Cleanup started
-2025-10-04 12:00:25.234 [INFO] Cleanup completed
-2025-10-04 12:00:30.567 [ERROR] Unexpected error occurred
-2025-10-04 12:01:00.890 [INFO] Application shutdown
-"""
+# ---------------------------------------------------------------------------
+# Dynamic test-data generators -- timestamps are always in the near future
+# so that loggrep's default "now" filter lets them through.
+# ---------------------------------------------------------------------------
 
-SAMPLE_LOG_ANDROID_LOGCAT = """10-04 11:59:59.123  1234  5678 I ActivityManager: Starting activity
-10-04 12:00:01.456  1234  5678 I ActivityManager: Activity started
-10-04 12:00:05.789  1234  5678 D NetworkManager: Network request sent
-10-04 12:00:10.012  1234  5678 E NetworkManager: Connection failed
-10-04 12:00:12.345  1234  5678 W NetworkManager: Retrying connection
-10-04 12:00:15.678  1234  5678 I NetworkManager: Connection established
-10-04 12:00:20.901  1234  5678 D LocationManager: GPS update received
-10-04 12:00:25.234  1234  5678 I LocationManager: Location updated
-10-04 12:00:30.567  1234  5678 E SystemManager: Low memory warning
-10-04 12:01:00.890  1234  5678 I ActivityManager: Activity stopped
-"""
+
+def _fmt_syslog(dt):
+    """Format datetime as syslog: 'Feb  4 12:00:00'."""
+    return f"{dt.strftime('%b')}{dt.day:>3} {dt.strftime('%H:%M:%S')}"
+
+
+def _fmt_iso8601(dt):
+    """Format datetime as ISO 8601: '2026-02-21 12:00:00.123'."""
+    return dt.strftime("%Y-%m-%d %H:%M:%S.") + f"{dt.microsecond // 1000:03d}"
+
+
+def _fmt_logcat(dt):
+    """Format datetime as Android logcat: '02-21 12:00:00.123'."""
+    return dt.strftime("%m-%d %H:%M:%S.") + f"{dt.microsecond // 1000:03d}"
+
+
+# Offsets (in seconds from base) shared by all three formats.
+_OFFSETS = [0, 2, 6, 11, 13, 16, 21, 26, 31, 61]
+
+
+def make_syslog_data(base_offset=60):
+    """Generate syslog-format log data with future timestamps.
+
+    Returns (data_string, base_time, list_of_datetimes).
+    """
+    base = datetime.now() + timedelta(seconds=base_offset)
+    times = [base + timedelta(seconds=s) for s in _OFFSETS]
+    lines = [
+        f"{_fmt_syslog(times[0])} server1 service[1234]: Starting service",
+        f"{_fmt_syslog(times[1])} server1 service[1234]: Service started successfully",
+        f"{_fmt_syslog(times[2])} server1 service[1234]: Processing request 1",
+        f"{_fmt_syslog(times[3])} server1 service[1234]: ERROR: Database connection failed",
+        f"{_fmt_syslog(times[4])} server1 service[1234]: WARN: Retrying database connection",
+        f"{_fmt_syslog(times[5])} server1 service[1234]: INFO: Database connection restored",
+        f"{_fmt_syslog(times[6])} server1 service[1234]: Processing request 2",
+        f"{_fmt_syslog(times[7])} server1 service[1234]: Request 2 completed successfully",
+        f"{_fmt_syslog(times[8])} server1 service[1234]: ERROR: Memory usage high",
+        f"{_fmt_syslog(times[9])} server1 service[1234]: Service stopped",
+    ]
+    return "\n".join(lines) + "\n", base, times
+
+
+def make_iso8601_data(base_offset=60):
+    """Generate ISO 8601 log data with future timestamps."""
+    base = datetime.now() + timedelta(seconds=base_offset)
+    times = [base + timedelta(seconds=s) for s in _OFFSETS]
+    lines = [
+        f"{_fmt_iso8601(times[0])} [INFO] Application starting",
+        f"{_fmt_iso8601(times[1])} [INFO] Configuration loaded",
+        f"{_fmt_iso8601(times[2])} [DEBUG] Processing user request",
+        f"{_fmt_iso8601(times[3])} [ERROR] Database connection timeout",
+        f"{_fmt_iso8601(times[4])} [WARN] Retrying operation",
+        f"{_fmt_iso8601(times[5])} [INFO] Operation completed",
+        f"{_fmt_iso8601(times[6])} [DEBUG] Cleanup started",
+        f"{_fmt_iso8601(times[7])} [INFO] Cleanup completed",
+        f"{_fmt_iso8601(times[8])} [ERROR] Unexpected error occurred",
+        f"{_fmt_iso8601(times[9])} [INFO] Application shutdown",
+    ]
+    return "\n".join(lines) + "\n", base, times
+
+
+def make_logcat_data(base_offset=60):
+    """Generate Android logcat log data with future timestamps."""
+    base = datetime.now() + timedelta(seconds=base_offset)
+    times = [base + timedelta(seconds=s) for s in _OFFSETS]
+    lines = [
+        f"{_fmt_logcat(times[0])}  1234  5678 I ActivityManager: Starting activity",
+        f"{_fmt_logcat(times[1])}  1234  5678 I ActivityManager: Activity started",
+        f"{_fmt_logcat(times[2])}  1234  5678 D NetworkManager: Network request sent",
+        f"{_fmt_logcat(times[3])}  1234  5678 E NetworkManager: Connection failed",
+        f"{_fmt_logcat(times[4])}  1234  5678 W NetworkManager: Retrying connection",
+        f"{_fmt_logcat(times[5])}  1234  5678 I NetworkManager: Connection established",
+        f"{_fmt_logcat(times[6])}  1234  5678 D LocationManager: GPS update received",
+        f"{_fmt_logcat(times[7])}  1234  5678 I LocationManager: Location updated",
+        f"{_fmt_logcat(times[8])}  1234  5678 E SystemManager: Low memory warning",
+        f"{_fmt_logcat(times[9])}  1234  5678 I ActivityManager: Activity stopped",
+    ]
+    return "\n".join(lines) + "\n", base, times
+
+
+# ---------------------------------------------------------------------------
+# Test helpers
+# ---------------------------------------------------------------------------
 
 
 def run_loggrep(args, input_data=None, expect_error=False):
     """Helper function to run loggrep with given arguments and input."""
-    # Use PYTHONPATH to ensure we can import our modules for coverage
-    import os
-    import sys
-
     env = os.environ.copy()
     env["PYTHONPATH"] = str(TEST_DIR / "src") + ":" + env.get("PYTHONPATH", "")
 
-    # On Windows, run with python explicitly to avoid executable issues
     if sys.platform == "win32":
         cmd = [sys.executable, str(LOGGREP_PATH)] + args
     else:
@@ -88,7 +136,8 @@ def run_loggrep(args, input_data=None, expect_error=False):
 
     if not expect_error and process.returncode != 0:
         pytest.fail(
-            f"loggrep failed with exit code {process.returncode}\nstderr: {process.stderr}\nstdout: {process.stdout}"
+            f"loggrep failed with exit code {process.returncode}\n"
+            f"stderr: {process.stderr}\nstdout: {process.stdout}"
         )
 
     return process
@@ -102,12 +151,18 @@ def create_temp_logfile(content):
     return temp_file.name
 
 
+# ===================================================================
+# Tests
+# ===================================================================
+
+
 class TestBasicFunctionality:
     """Test basic search functionality."""
 
     def test_basic_pattern_search(self):
         """Test basic pattern search in log file."""
-        temp_file = create_temp_logfile(SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        temp_file = create_temp_logfile(data)
         try:
             result = run_loggrep(["ERROR", "--file", temp_file])
             assert "ERROR: Database connection failed" in result.stdout
@@ -118,22 +173,24 @@ class TestBasicFunctionality:
 
     def test_stdin_input(self):
         """Test reading from stdin."""
-        result = run_loggrep(["ERROR"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["ERROR"], input_data=data)
         assert "ERROR: Database connection failed" in result.stdout
         assert "ERROR: Memory usage high" in result.stdout
 
     def test_no_matches(self):
         """Test when pattern doesn't match anything."""
-        result = run_loggrep(["NONEXISTENT"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["NONEXISTENT"], input_data=data)
         assert result.stdout.strip() == ""
 
     def test_multiple_patterns(self):
         """Test searching for multiple patterns (OR logic)."""
-        result = run_loggrep(["ERROR", "WARN"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["ERROR", "WARN"], input_data=data)
         assert "ERROR: Database connection failed" in result.stdout
         assert "ERROR: Memory usage high" in result.stdout
         assert "WARN: Retrying database connection" in result.stdout
-        # Should have 3 total matches
         assert len(result.stdout.strip().split("\n")) == 3
 
 
@@ -142,16 +199,16 @@ class TestRegexSupport:
 
     def test_regex_pattern(self):
         """Test using regex patterns."""
-        # Test regex pattern for ERROR or WARN
-        result = run_loggrep([r"(ERROR|WARN)"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        result = run_loggrep([r"(ERROR|WARN)"], input_data=data)
         assert "ERROR: Database connection failed" in result.stdout
         assert "WARN: Retrying database connection" in result.stdout
         assert "ERROR: Memory usage high" in result.stdout
 
     def test_regex_word_boundaries(self):
         """Test regex with word boundaries."""
-        result = run_loggrep([r"\bservice\b"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
-        # Should match lines with "service" as a whole word
+        data, _, _ = make_syslog_data()
+        result = run_loggrep([r"\bservice\b"], input_data=data)
         lines = result.stdout.strip().split("\n")
         assert len(lines) > 0
         for line in lines:
@@ -159,9 +216,8 @@ class TestRegexSupport:
 
     def test_invalid_regex(self):
         """Test handling of invalid regex patterns."""
-        result = run_loggrep(
-            ["[invalid"], input_data=SAMPLE_LOG_UNIX_SYSLOG, expect_error=True
-        )
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["[invalid"], input_data=data, expect_error=True)
         assert result.returncode != 0
         assert "Invalid regex pattern" in result.stderr
 
@@ -171,16 +227,18 @@ class TestCaseInsensitive:
 
     def test_case_insensitive_flag(self):
         """Test -i/--ignore-case flag."""
-        result = run_loggrep(["-i", "error"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["-i", "error"], input_data=data)
         assert "ERROR: Database connection failed" in result.stdout
         assert "ERROR: Memory usage high" in result.stdout
 
     def test_case_sensitive_default(self):
         """Test that search is case-sensitive by default."""
-        result = run_loggrep(["error"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
-        assert result.stdout.strip() == ""  # Should find nothing with lowercase "error"
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["error"], input_data=data)
+        assert result.stdout.strip() == ""
 
-        result = run_loggrep(["ERROR"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
+        result = run_loggrep(["ERROR"], input_data=data)
         assert "ERROR: Database connection failed" in result.stdout
 
 
@@ -189,22 +247,21 @@ class TestInvertMatch:
 
     def test_invert_match(self):
         """Test -v/--invert-match flag."""
-        result = run_loggrep(["-v", "ERROR"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
-        # Should get all lines except those containing ERROR
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["-v", "ERROR"], input_data=data)
         lines = result.stdout.strip().split("\n")
         for line in lines:
             assert "ERROR" not in line
-        # Should have 8 lines (10 total - 2 ERROR lines)
         assert len(lines) == 8
 
     def test_invert_match_with_multiple_patterns(self):
         """Test invert match with multiple patterns."""
-        result = run_loggrep(["-v", "ERROR", "WARN"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["-v", "ERROR", "WARN"], input_data=data)
         lines = result.stdout.strip().split("\n")
         for line in lines:
             assert "ERROR" not in line
             assert "WARN" not in line
-        # Should have 7 lines (10 total - 2 ERROR - 1 WARN)
         assert len(lines) == 7
 
 
@@ -213,33 +270,24 @@ class TestContextLines:
 
     def test_after_context(self):
         """Test -A flag (lines after match)."""
-        result = run_loggrep(
-            ["-A", "2", "ERROR.*Database"], input_data=SAMPLE_LOG_UNIX_SYSLOG
-        )
-        lines = result.stdout.strip().split("\n")
-        # Should include the match line plus 2 lines after
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["-A", "2", "ERROR.*Database"], input_data=data)
         assert "ERROR: Database connection failed" in result.stdout
         assert "WARN: Retrying database connection" in result.stdout
         assert "INFO: Database connection restored" in result.stdout
 
     def test_before_context(self):
         """Test -B flag (lines before match)."""
-        result = run_loggrep(
-            ["-B", "2", "ERROR.*Database"], input_data=SAMPLE_LOG_UNIX_SYSLOG
-        )
-        lines = result.stdout.strip().split("\n")
-        # Should include 2 lines before plus the match line
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["-B", "2", "ERROR.*Database"], input_data=data)
         assert "Service started successfully" in result.stdout
         assert "Processing request 1" in result.stdout
         assert "ERROR: Database connection failed" in result.stdout
 
     def test_context_around(self):
         """Test -C flag (lines around match)."""
-        result = run_loggrep(
-            ["-C", "1", "ERROR.*Database"], input_data=SAMPLE_LOG_UNIX_SYSLOG
-        )
-        lines = result.stdout.strip().split("\n")
-        # Should include 1 line before, the match, and 1 line after
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["-C", "1", "ERROR.*Database"], input_data=data)
         assert "Processing request 1" in result.stdout
         assert "ERROR: Database connection failed" in result.stdout
         assert "WARN: Retrying database connection" in result.stdout
@@ -250,13 +298,14 @@ class TestTimestampParsing:
 
     def test_unix_syslog_timestamps(self):
         """Test parsing Unix syslog timestamp format."""
-        temp_file = create_temp_logfile(SAMPLE_LOG_UNIX_SYSLOG)
+        data, base, times = make_syslog_data()
+        temp_file = create_temp_logfile(data)
         try:
-            # Search after 12:00:15
+            # Startup between entry 5 (+16s) and entry 6 (+21s)
+            startup = _fmt_syslog(base + timedelta(seconds=18))
             result = run_loggrep(
-                ["Processing", "--file", temp_file, "--startup-time", "Oct 4 12:00:15"]
+                ["Processing", "--file", temp_file, "--startup-time", startup]
             )
-            # Should only find "Processing request 2" which is at 12:00:20
             assert "Processing request 2" in result.stdout
             assert "Processing request 1" not in result.stdout
         finally:
@@ -264,17 +313,17 @@ class TestTimestampParsing:
 
     def test_iso8601_timestamps(self):
         """Test parsing ISO 8601 timestamp format."""
-        temp_file = create_temp_logfile(SAMPLE_LOG_ISO8601)
+        data, base, times = make_iso8601_data()
+        temp_file = create_temp_logfile(data)
         try:
-            # Search after 12:00:15
+            # Startup between entry 4 (+13s) and entry 5 (+16s)
+            startup = _fmt_iso8601(base + timedelta(seconds=15))
             result = run_loggrep(
-                ["INFO", "--file", temp_file, "--startup-time", "2025-10-04 12:00:15"]
+                ["INFO", "--file", temp_file, "--startup-time", startup]
             )
-            # Should find INFO messages after 12:00:15
             assert "Operation completed" in result.stdout
             assert "Cleanup completed" in result.stdout
             assert "Application shutdown" in result.stdout
-            # Should not find earlier INFO messages
             assert "Application starting" not in result.stdout
             assert "Configuration loaded" not in result.stdout
         finally:
@@ -282,90 +331,73 @@ class TestTimestampParsing:
 
     def test_android_logcat_timestamps(self):
         """Test parsing Android logcat timestamp format."""
-        temp_file = create_temp_logfile(SAMPLE_LOG_ANDROID_LOGCAT)
+        data, base, times = make_logcat_data()
+        temp_file = create_temp_logfile(data)
         try:
-            # Search after 12:00:20
+            # Startup between entry 7 (+26s) and entry 8 (+31s)
+            startup = _fmt_logcat(base + timedelta(seconds=28))
             result = run_loggrep(
-                [
-                    "ActivityManager",
-                    "--file",
-                    temp_file,
-                    "--startup-time",
-                    "10-04 12:00:20",
-                ]
+                ["ActivityManager", "--file", temp_file, "--startup-time", startup]
             )
-            # Should only find the last ActivityManager message
             assert "Activity stopped" in result.stdout
             assert "Starting activity" not in result.stdout
             assert "Activity started" not in result.stdout
         finally:
             os.unlink(temp_file)
 
-    def test_no_startup_time_uses_first_timestamp(self):
-        """Test that without startup-time, first timestamp is used for files."""
-        temp_file = create_temp_logfile(SAMPLE_LOG_UNIX_SYSLOG)
+    def test_no_startup_time_shows_future_entries(self):
+        """Without --startup-time, file mode defaults to now and shows future entries."""
+        data, _, _ = make_syslog_data()
+        temp_file = create_temp_logfile(data)
         try:
             result = run_loggrep(["service", "--file", temp_file])
-            # Should find all lines with "service" since we start from first timestamp
             lines = result.stdout.strip().split("\n")
-            assert len(lines) > 5  # Should find multiple service-related lines
+            assert len(lines) > 5
         finally:
             os.unlink(temp_file)
 
     def test_stdin_defaults_to_current_time(self):
-        """Test that stdin input with --live flag defaults to current time."""
-        from datetime import datetime, timedelta
-
-        # Create log data with timestamps - some old, some very recent
+        """Test that stdin filters old entries and shows future ones."""
         now = datetime.now()
         old_time = now - timedelta(hours=1)
-        recent_time = now + timedelta(
-            seconds=5
-        )  # Several seconds in future to ensure it's processed
+        future_time = now + timedelta(seconds=60)
 
-        log_data = f"""{old_time.strftime('%Y-%m-%d %H:%M:%S')} [INFO] Old message should be filtered
-{recent_time.strftime('%Y-%m-%d %H:%M:%S')} [ERROR] Recent message should appear
-"""
+        log_data = (
+            f"{old_time.strftime('%Y-%m-%d %H:%M:%S')} [INFO] Old message should be filtered\n"
+            f"{future_time.strftime('%Y-%m-%d %H:%M:%S')} [ERROR] Recent message should appear\n"
+        )
 
-        result = run_loggrep(["ERROR", "--live"], input_data=log_data)
-        # Should only find the recent ERROR message, not the old one
+        result = run_loggrep(["ERROR"], input_data=log_data)
         assert "Recent message should appear" in result.stdout
         assert "Old message should be filtered" not in result.stdout
 
-    def test_file_uses_first_timestamp_not_current_time(self):
-        """Test that file input uses first timestamp, not current time."""
-        from datetime import datetime, timedelta
-
-        # Create log data with all timestamps in the past
+    def test_file_defaults_to_now_filters_past(self):
+        """Test that file mode filters out past entries by default."""
         past_time = datetime.now() - timedelta(hours=2)
         past_time2 = past_time + timedelta(minutes=5)
 
-        log_data = f"""{past_time.strftime('%Y-%m-%d %H:%M:%S')} [INFO] First message
-{past_time2.strftime('%Y-%m-%d %H:%M:%S')} [ERROR] Second message
-"""
+        log_data = (
+            f"{past_time.strftime('%Y-%m-%d %H:%M:%S')} [INFO] First message\n"
+            f"{past_time2.strftime('%Y-%m-%d %H:%M:%S')} [ERROR] Second message\n"
+        )
 
         temp_file = create_temp_logfile(log_data)
         try:
+            # Default file mode filters to "now" -- past logs excluded
             result = run_loggrep(["ERROR", "--file", temp_file])
-            # Should find the ERROR message even though it's in the past
-            # because file mode uses first timestamp, not current time
+            assert "Second message" not in result.stdout
+
+            # --no-live disables filtering
+            result = run_loggrep(["ERROR", "--no-live", "--file", temp_file])
             assert "Second message" in result.stdout
         finally:
             os.unlink(temp_file)
 
     def test_live_flag_functionality(self):
         """Test that --live flag works for real-time log streaming simulation."""
-        import subprocess
-        import sys
-        import threading
-        import time
-        from datetime import datetime
-
-        # Use async bash to simulate live log streaming
         env = os.environ.copy()
         env["PYTHONPATH"] = str(TEST_DIR / "src") + ":" + env.get("PYTHONPATH", "")
 
-        # Build command with Windows compatibility
         if sys.platform == "win32":
             cmd = [sys.executable, str(LOGGREP_PATH), "ERROR", "--live"]
         else:
@@ -382,23 +414,20 @@ class TestTimestampParsing:
         )
 
         try:
-            # Generate log entries in real-time
             now = datetime.now()
+            old_entry = (
+                f"{(now - timedelta(hours=1)).strftime('%Y-%m-%d %H:%M:%S')}"
+                " [ERROR] Old error message\n"
+            )
+            future_entry = (
+                f"{(now + timedelta(seconds=60)).strftime('%Y-%m-%d %H:%M:%S')}"
+                " [ERROR] Future error message\n"
+            )
 
-            # Send an old log entry (should be filtered out)
-            old_entry = f"{(now - timedelta(hours=1)).strftime('%Y-%m-%d %H:%M:%S')} [ERROR] Old error message\n"
+            stdout, stderr = process.communicate(
+                input=old_entry + future_entry, timeout=5
+            )
 
-            # Send a future log entry (should appear - this simulates live streaming)
-            future_entry = f"{(now + timedelta(seconds=2)).strftime('%Y-%m-%d %H:%M:%S')} [ERROR] Future error message\n"
-
-            # Combine all input
-            input_data = old_entry + future_entry
-
-            # Send all input and get output
-            stdout, stderr = process.communicate(input=input_data, timeout=5)
-
-            # Verify that only future entries appear (old ones filtered out)
-            # This is the expected behavior for live mode - it filters based on startup time
             assert "Future error message" in stdout
             assert "Old error message" not in stdout
 
@@ -413,15 +442,9 @@ class TestTimestampParsing:
 
     def test_live_streaming_simulation(self):
         """Test live streaming like 'tail -f' or 'adb logcat' scenarios."""
-        import subprocess
-        import sys
-        import time
-        from datetime import datetime
-
         env = os.environ.copy()
         env["PYTHONPATH"] = str(TEST_DIR / "src") + ":" + env.get("PYTHONPATH", "")
 
-        # Build command with Windows compatibility
         if sys.platform == "win32":
             cmd = [sys.executable, str(LOGGREP_PATH), "ActivityManager", "--live"]
         else:
@@ -440,14 +463,13 @@ class TestTimestampParsing:
         output_lines = []
 
         def read_output():
-            """Read output from process in separate thread."""
             try:
                 while True:
                     line = process.stdout.readline()
                     if not line:
                         break
                     output_lines.append(line.strip())
-            except:
+            except Exception:
                 pass
 
         output_thread = threading.Thread(target=read_output, daemon=True)
@@ -455,44 +477,34 @@ class TestTimestampParsing:
 
         try:
             now = datetime.now()
-
-            # Simulate old log entries (should be filtered)
             old_entries = [
-                f"{(now - timedelta(minutes=30)).strftime('%m-%d %H:%M:%S.%f')[:-3]}  1234  5678 I ActivityManager: Old activity start",
-                f"{(now - timedelta(minutes=20)).strftime('%m-%d %H:%M:%S.%f')[:-3]}  1234  5678 I ActivityManager: Old activity pause",
+                f"{(now - timedelta(minutes=30)).strftime('%m-%d %H:%M:%S.%f')[:-3]}"
+                "  1234  5678 I ActivityManager: Old activity start",
+                f"{(now - timedelta(minutes=20)).strftime('%m-%d %H:%M:%S.%f')[:-3]}"
+                "  1234  5678 I ActivityManager: Old activity pause",
             ]
-
-            # Simulate current/recent log entries (should appear)
             future_entries = [
-                f"{(now + timedelta(seconds=1)).strftime('%m-%d %H:%M:%S.%f')[:-3]}  1234  5678 I ActivityManager: Future activity launch",
-                f"{(now + timedelta(seconds=2)).strftime('%m-%d %H:%M:%S.%f')[:-3]}  1234  5678 I ActivityManager: Recent activity resume",
+                f"{(now + timedelta(seconds=60)).strftime('%m-%d %H:%M:%S.%f')[:-3]}"
+                "  1234  5678 I ActivityManager: Future activity launch",
+                f"{(now + timedelta(seconds=61)).strftime('%m-%d %H:%M:%S.%f')[:-3]}"
+                "  1234  5678 I ActivityManager: Recent activity resume",
             ]
 
-            # Send old entries first
             for entry in old_entries:
                 process.stdin.write(entry + "\n")
                 process.stdin.flush()
-                time.sleep(0.05)  # Small delay to simulate real streaming
+                time.sleep(0.05)
 
-            # Send current entries
             for entry in future_entries:
                 process.stdin.write(entry + "\n")
                 process.stdin.flush()
                 time.sleep(0.05)
 
-            # Give some time for processing
             time.sleep(0.2)
-
-            # Close stdin to signal end
             process.stdin.close()
-
-            # Wait for process to finish
             process.wait(timeout=3)
 
-            # Check results
             output_text = "\n".join(output_lines)
-
-            # Should show current activities but not old ones
             assert "Future activity launch" in output_text
             assert "Recent activity resume" in output_text
             assert "Old activity start" not in output_text
@@ -512,27 +524,20 @@ class TestColorOutput:
 
     def test_color_auto_with_tty(self):
         """Test color=auto behavior (default)."""
-        # When output is not a TTY (like in tests), color should be disabled by default
-        result = run_loggrep(["ERROR"], input_data=SAMPLE_LOG_UNIX_SYSLOG)
-        # Should not contain ANSI color codes when output is not a TTY
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["ERROR"], input_data=data)
         assert "\x1b[" not in result.stdout
 
     def test_color_always(self):
         """Test --color=always flag."""
-        result = run_loggrep(
-            ["--color=always", "ERROR"], input_data=SAMPLE_LOG_UNIX_SYSLOG
-        )
-        # Should contain ANSI color codes when forced
-        if "colorama" in sys.modules or True:  # colorama should be available
-            # The exact color codes might vary, but should have some coloring
-            assert "ERROR" in result.stdout
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["--color=always", "ERROR"], input_data=data)
+        assert "ERROR" in result.stdout
 
     def test_color_never(self):
         """Test --color=never flag."""
-        result = run_loggrep(
-            ["--color=never", "ERROR"], input_data=SAMPLE_LOG_UNIX_SYSLOG
-        )
-        # Should not contain ANSI color codes
+        data, _, _ = make_syslog_data()
+        result = run_loggrep(["--color=never", "ERROR"], input_data=data)
         assert "\x1b[" not in result.stdout
         assert "ERROR" in result.stdout
 
@@ -559,29 +564,41 @@ class TestEdgeCases:
 
     def test_log_without_timestamps(self):
         """Test handling of log without recognizable timestamps."""
-        log_without_timestamps = """Line 1 without timestamp
-Line 2 without timestamp
-ERROR: Something went wrong
-Line 4 without timestamp
-"""
-        result = run_loggrep(["ERROR"], input_data=log_without_timestamps)
-        # Should still work, but might not apply timestamp filtering
+        log_without_timestamps = (
+            "Line 1 without timestamp\n"
+            "Line 2 without timestamp\n"
+            "ERROR: Something went wrong\n"
+            "Line 4 without timestamp\n"
+        )
+        # --no-live is appropriate here: no timestamps in data
+        result = run_loggrep(["ERROR", "--no-live"], input_data=log_without_timestamps)
         assert "ERROR: Something went wrong" in result.stdout
 
     def test_invalid_startup_time(self):
         """Test handling of invalid startup time format."""
-        temp_file = create_temp_logfile(SAMPLE_LOG_UNIX_SYSLOG)
+        data, _, _ = make_syslog_data()
+        temp_file = create_temp_logfile(data)
         try:
             result = run_loggrep(
                 ["ERROR", "--file", temp_file, "--startup-time", "invalid-time"]
             )
-            # Should handle gracefully (might default to first timestamp or show error)
-            # The exact behavior depends on implementation
-            assert (
-                result.returncode == 0 or result.returncode != 0
-            )  # Either way is acceptable
+            assert result.returncode == 0 or result.returncode != 0
         finally:
             os.unlink(temp_file)
+
+    def test_live_mode_without_timestamps(self):
+        """Test that the no-timestamp fallback kicks in for stdin."""
+        log_without_timestamps = (
+            "Line 1 without timestamp\n"
+            "Line 2 containing walkie\n"
+            "Line 3 without timestamp\n"
+            "Line 4 containing walkie again\n"
+            "Line 5 without timestamp\n"
+        )
+        result = run_loggrep(["walkie"], input_data=log_without_timestamps)
+        output_lines = result.stdout.strip().split("\n")
+        assert len(output_lines) >= 1
+        assert "Line 4 containing walkie again" in result.stdout
 
 
 class TestIntegrationScenarios:
@@ -589,9 +606,11 @@ class TestIntegrationScenarios:
 
     def test_complex_search_scenario(self):
         """Test a complex search combining multiple features."""
-        temp_file = create_temp_logfile(SAMPLE_LOG_UNIX_SYSLOG)
+        data, base, times = make_syslog_data()
+        temp_file = create_temp_logfile(data)
         try:
-            # Case-insensitive search for error/warn after 12:00:05 with context
+            # Startup between entry 2 (+6s) and entry 3 (+11s)
+            startup = _fmt_syslog(base + timedelta(seconds=8))
             result = run_loggrep(
                 [
                     "-i",
@@ -602,12 +621,10 @@ class TestIntegrationScenarios:
                     "--file",
                     temp_file,
                     "--startup-time",
-                    "Oct 4 12:00:05",
+                    startup,
                 ]
             )
-
             lines = result.stdout.strip().split("\n")
-            # Should include context lines around matches
             assert len(lines) > 2
             assert any("ERROR" in line for line in lines)
             assert any("WARN" in line for line in lines)
@@ -616,16 +633,18 @@ class TestIntegrationScenarios:
 
     def test_real_world_log_patterns(self):
         """Test patterns that would be used in real-world scenarios."""
-        # Test searching for specific error codes, IPs, etc.
-        log_with_patterns = """Oct  4 12:00:01 server1 nginx[1234]: 192.168.1.100 - GET /api/v1/users
-Oct  4 12:00:02 server1 nginx[1234]: 192.168.1.101 - POST /api/v1/login
-Oct  4 12:00:03 server1 nginx[1234]: 192.168.1.100 - GET /api/v1/users - 200 OK
-Oct  4 12:00:04 server1 nginx[1234]: 192.168.1.101 - POST /api/v1/login - 401 Unauthorized
-Oct  4 12:00:05 server1 nginx[1234]: 192.168.1.102 - GET /api/v1/admin - 403 Forbidden
-Oct  4 12:00:06 server1 nginx[1234]: 192.168.1.100 - GET /api/v1/data - 500 Internal Server Error
-"""
+        base = datetime.now() + timedelta(seconds=60)
+        offsets = [0, 1, 2, 3, 4, 5]
+        times = [base + timedelta(seconds=s) for s in offsets]
+        log_with_patterns = (
+            f"{_fmt_syslog(times[0])} server1 nginx[1234]: 192.168.1.100 - GET /api/v1/users\n"
+            f"{_fmt_syslog(times[1])} server1 nginx[1234]: 192.168.1.101 - POST /api/v1/login\n"
+            f"{_fmt_syslog(times[2])} server1 nginx[1234]: 192.168.1.100 - GET /api/v1/users - 200 OK\n"
+            f"{_fmt_syslog(times[3])} server1 nginx[1234]: 192.168.1.101 - POST /api/v1/login - 401 Unauthorized\n"
+            f"{_fmt_syslog(times[4])} server1 nginx[1234]: 192.168.1.102 - GET /api/v1/admin - 403 Forbidden\n"
+            f"{_fmt_syslog(times[5])} server1 nginx[1234]: 192.168.1.100 - GET /api/v1/data - 500 Internal Server Error\n"
+        )
 
-        # Test searching for specific HTTP status codes
         result = run_loggrep([r"\b(401|403|500)\b"], input_data=log_with_patterns)
         lines = result.stdout.strip().split("\n")
         assert len(lines) == 3
@@ -633,15 +652,13 @@ Oct  4 12:00:06 server1 nginx[1234]: 192.168.1.100 - GET /api/v1/data - 500 Inte
         assert "403 Forbidden" in result.stdout
         assert "500 Internal Server Error" in result.stdout
 
-        # Test searching for specific IP addresses
         result = run_loggrep([r"192\.168\.1\.100"], input_data=log_with_patterns)
         lines = result.stdout.strip().split("\n")
         assert len(lines) == 3
 
-        # Test searching for API endpoints
         result = run_loggrep([r"/api/v1/\w+"], input_data=log_with_patterns)
         lines = result.stdout.strip().split("\n")
-        assert len(lines) == 6  # All lines should match
+        assert len(lines) == 6
 
 
 def test_help_option():
@@ -655,9 +672,6 @@ def test_help_option():
 
 def test_version_compatibility():
     """Test that the script runs with the documented Python version."""
-    # The README specifies Python 3.7+, ensure we're compatible
-    import sys
-
     assert sys.version_info >= (3, 7), "Python 3.7+ required as per README"
 
 
@@ -673,55 +687,33 @@ class TestErrorHandling:
 
     def test_permission_denied_error(self):
         """Test handling of permission denied on file."""
-        import os
-        import stat
-        import sys
-        import tempfile
-
-        # Skip on Windows as it doesn't have getuid and permission model is different
         if sys.platform == "win32":
-            import pytest
-
             pytest.skip("Permission test not applicable on Windows")
 
-        # Create a temporary file and make it unreadable
         with tempfile.NamedTemporaryFile(mode="w", delete=False) as f:
             f.write("test content\n")
             temp_file = f.name
 
         try:
-            # Remove read permissions - use different approach for Docker/root
-            if (
-                hasattr(os, "getuid") and os.getuid() == 0
-            ):  # Running as root (common in Docker)
-                # Skip this test when running as root since root can read any file
-                import pytest
-
+            if hasattr(os, "getuid") and os.getuid() == 0:
                 pytest.skip("Permission test not applicable when running as root")
             else:
-                os.chmod(temp_file, 0o200)  # Write only, no read
-
+                os.chmod(temp_file, 0o200)
                 result = run_loggrep(["test", "--file", temp_file], expect_error=True)
                 assert result.returncode == 2
                 assert "Permission denied" in result.stderr
         finally:
-            # Restore permissions and cleanup
             try:
                 os.chmod(temp_file, 0o644)
                 os.unlink(temp_file)
-            except:
+            except Exception:
                 pass
 
     def test_directory_error(self):
         """Test handling when file argument is a directory."""
-        import sys
-        import tempfile
-
-        # Use a temporary directory
         with tempfile.TemporaryDirectory() as temp_dir:
             result = run_loggrep(["test", "--file", temp_dir], expect_error=True)
             assert result.returncode == 2
-            # Windows shows "Permission denied" for directories, Unix shows "Is a directory"
             if sys.platform == "win32":
                 assert "Permission denied" in result.stderr
             else:
@@ -740,14 +732,8 @@ class TestErrorHandling:
 
     def test_broken_pipe_handling(self):
         """Test graceful handling of broken pipe."""
-        import os
-        import subprocess
-        import sys
-
-        # Create a large file
         temp_file = create_temp_logfile("line with content\n" * 1000)
         try:
-            # Build command with Windows compatibility
             if sys.platform == "win32":
                 cmd = [
                     sys.executable,
@@ -759,50 +745,35 @@ class TestErrorHandling:
             else:
                 cmd = [str(LOGGREP_PATH), "content", "--file", temp_file]
 
-            # Start loggrep process
             process = subprocess.Popen(
                 cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True
             )
-
-            # Read just a bit and then close to simulate broken pipe
             if process.stdout:
                 process.stdout.read(100)
                 process.stdout.close()
-
-            # Wait for process to complete
             process.wait()
-
-            # Should handle broken pipe gracefully (exit code 0)
-            # Note: broken pipe might return various exit codes depending on OS
-            assert process.returncode in [0, 1, 120]  # Allow for different OS behaviors
+            assert process.returncode in [0, 1, 120]
         finally:
             try:
                 os.unlink(temp_file)
-            except:
+            except Exception:
                 pass
 
     def test_invalid_regex_error_handling(self):
         """Test handling of invalid regex patterns."""
-        # Test with an invalid regex that will cause an error
-        log_data = "2023-10-04 12:00:00 Test message\n"
-
-        # Invalid regex with unmatched bracket
-        result = run_loggrep(["[invalid"], input_data=log_data, expect_error=True)
+        data, _, _ = make_iso8601_data()
+        result = run_loggrep(["[invalid"], input_data=data, expect_error=True)
         assert result.returncode == 1
-        assert result.stdout.strip() == ""  # No matches due to error
+        assert result.stdout.strip() == ""
 
     def test_colorama_import_scenarios(self):
         """Test scenarios that exercise colorama import handling."""
-        # Test color detection works regardless of colorama availability
         result = run_loggrep(["--help"])
         assert result.returncode == 0
         assert "color" in result.stdout.lower()
 
-        # Test that auto color detection doesn't crash even if colorama fails
-        # This exercises the ImportError handling in _supports_color()
-        result = run_loggrep(
-            ["INFO", "--color", "auto"], input_data="2023-10-04 12:00:00 [INFO] Test\n"
-        )
+        data, _, _ = make_iso8601_data()
+        result = run_loggrep(["INFO", "--color", "auto"], input_data=data)
         assert result.returncode == 0
 
 
@@ -811,21 +782,15 @@ class TestColorHandling:
 
     def test_color_auto_without_tty(self):
         """Test color=auto behavior when not in TTY."""
-        # When running in subprocess (not TTY), should not use colors
-        result = run_loggrep(
-            ["INFO", "--color", "auto"],
-            input_data="2023-10-04 12:00:00 [INFO] Test message\n",
-        )
+        data, _, _ = make_iso8601_data()
+        result = run_loggrep(["INFO", "--color", "auto"], input_data=data)
         assert result.returncode == 0
-        # Output should not contain ANSI color codes when not in TTY
         assert "\033[" not in result.stdout
 
     def test_color_never(self):
         """Test that --color never disables colors."""
-        result = run_loggrep(
-            ["INFO", "--color", "never"],
-            input_data="2023-10-04 12:00:00 [INFO] Test message\n",
-        )
+        data, _, _ = make_iso8601_data()
+        result = run_loggrep(["INFO", "--color", "never"], input_data=data)
         assert result.returncode == 0
         assert "\033[" not in result.stdout
 
@@ -835,22 +800,25 @@ class TestMemoryAndPerformance:
 
     def test_large_input_handling(self):
         """Test handling of large input data."""
-        # Create a moderately large input to test memory handling
-        large_log = "2023-10-04 12:00:00 Test line\n" * 10000
+        future_time = (datetime.now() + timedelta(seconds=60)).strftime(
+            "%Y-%m-%d %H:%M:%S"
+        )
+        large_log = f"{future_time} Test line\n" * 10000
 
         result = run_loggrep(["Test"], input_data=large_log)
         assert result.returncode == 0
-        # Should find all matching lines
         assert result.stdout.count("Test line") == 10000
 
     def test_no_matches_large_file(self):
         """Test behavior with large file that has no matches."""
-        large_log = "2023-10-04 12:00:00 Other content\n" * 5000
+        future_time = (datetime.now() + timedelta(seconds=60)).strftime(
+            "%Y-%m-%d %H:%M:%S"
+        )
+        large_log = f"{future_time} Other content\n" * 5000
 
         result = run_loggrep(["NonExistent"], input_data=large_log)
-        # When no matches are found, loggrep returns 0 (success) but empty output
-        assert result.returncode == 0  # Process completed successfully
-        assert result.stdout.strip() == ""  # No output for no matches
+        assert result.returncode == 0
+        assert result.stdout.strip() == ""
 
 
 class TestRegexEdgeCases:
@@ -858,32 +826,86 @@ class TestRegexEdgeCases:
 
     def test_regex_special_characters(self):
         """Test regex with special characters."""
-        log_data = """2023-10-04 12:00:00 Price: $100.50
-2023-10-04 12:00:01 Email: user@domain.com
-2023-10-04 12:00:02 Path: /var/log/app.log
-2023-10-04 12:00:03 Query: SELECT * FROM users
-"""
+        future_time = (datetime.now() + timedelta(seconds=60)).strftime(
+            "%Y-%m-%d %H:%M:%S"
+        )
+        log_data = (
+            f"{future_time} Price: $100.50\n"
+            f"{future_time} Email: user@domain.com\n"
+            f"{future_time} Path: /var/log/app.log\n"
+            f"{future_time} Query: SELECT * FROM users\n"
+        )
 
-        # Test escaping special regex characters
         result = run_loggrep([r"\$\d+\.\d+"], input_data=log_data)
         assert "Price: $100.50" in result.stdout
 
-        # Test email pattern
         result = run_loggrep([r"\w+@\w+\.\w+"], input_data=log_data)
         assert "user@domain.com" in result.stdout
 
     def test_complex_regex_patterns(self):
         """Test complex regex patterns."""
-        log_data = """2023-10-04 12:00:00 [ERROR] Database connection failed: timeout after 30s
-2023-10-04 12:00:01 [WARN] Memory usage: 85%
-2023-10-04 12:00:02 [INFO] User login: john.doe
-2023-10-04 12:00:03 [ERROR] API error: 404 Not Found
-"""
+        future_time = (datetime.now() + timedelta(seconds=60)).strftime(
+            "%Y-%m-%d %H:%M:%S"
+        )
+        log_data = (
+            f"{future_time} [ERROR] Database connection failed: timeout after 30s\n"
+            f"{future_time} [WARN] Memory usage: 85%\n"
+            f"{future_time} [INFO] User login: john.doe\n"
+            f"{future_time} [ERROR] API error: 404 Not Found\n"
+        )
 
-        # Test lookahead/lookbehind if supported
         result = run_loggrep([r"ERROR.*\d+"], input_data=log_data)
         lines = result.stdout.strip().split("\n")
-        assert len(lines) == 2  # Two ERROR lines with numbers
+        assert len(lines) == 2
+
+
+class TestEncodingHandling:
+    """Test handling of various text encodings and binary data."""
+
+    def test_utf8_decode_errors_handling(self):
+        """Test that binary data doesn't cause UTF-8 decode errors."""
+        now = datetime.now()
+        current_date = now.strftime("%m-%d")
+        current_time = now.strftime("%H:%M:%S.%f")[:-3]
+
+        binary_data_script = f"""import sys
+import time
+sys.stdout.buffer.write(b'{current_date} {current_time} I walkie: Starting app\\n')
+time.sleep(0.001)
+sys.stdout.buffer.write(b'{current_date} {current_time} I walkie: \\xc0\\x80Invalid UTF-8 data\\n')
+time.sleep(0.001)
+sys.stdout.buffer.write(b'{current_date} {current_time} I walkie: Normal log line\\n')
+time.sleep(0.001)
+sys.stdout.buffer.write(b'{current_date} {current_time} E other: \\xff\\xfeMore binary data\\n')
+"""
+
+        process1 = subprocess.Popen(
+            [sys.executable, "-c", binary_data_script],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+
+        loggrep_path = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), "loggrep"
+        )
+
+        process2 = subprocess.Popen(
+            [sys.executable, loggrep_path, "walkie", "--no-live"],
+            stdin=process1.stdout,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            text=True,
+        )
+
+        process1.stdout.close()
+        stdout, stderr = process2.communicate()
+
+        assert "Unable to decode" not in stderr
+        assert "utf-8 codec can't decode" not in stderr
+        assert "walkie: Starting app" in stdout
+        assert "walkie:" in stdout
+        assert "walkie: Normal log line" in stdout
+        assert process2.returncode == 0
 
 
 if __name__ == "__main__":
